WRI-65: Writed one more test. Connected H2 database for tests
